+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Version NEXT
 ==============
 
-* Cart object is not saved to database if it is empty (#147)
+* Added get_product_reference method to Product (for extensibility)
 * Changed spelling mistakes in methods from `payed` to `paid` on the Order 
   model and on the API. This is potentially not backwards compatible in some 
   border cases.